Add available and state to SamsungTV remote
